Improve styles

+++ src/lib/builder/components/Field/parts/Images.svelte
@@ -22,39 +22,38 @@
 	}
 </script>
 
-<div
-	class="
-		flex gap-2
-		py-1 px-2
-		bg-neutral-700
-		border-neutral-600
-		border
-		rounded
-	"
->
+<div class="field">
 	<div class="flex flex-col w-full">
 		<input
-			id={name}
-			name={name}
+			id={$name}
 			type="file"
 			accept="image/png, image/gif, image/jpeg"
 			multiple
-			class="block w-full my-1"
+			class="block w-full file-btn file:p-2"
 			bind:this={el}
 			bind:files={inputFiles}
 		/>
 
-		{#each $files as file, index}
-			<div class="flex gap-2 my-1">
-				<div class="w-full overflow-hidden text-ellipsis">
-					{file.file.name}
-				</div>
-				<button
-					on:click={() => {removeFile(index);}}
-				>
-					<Icon src={X} class="w-6 h-6" />
-				</button>
+		{#if $files.length > 0}
+			<div class="mt-2">
+				{#each $files as file, index}
+					<div class="
+						py-1 px-2
+						border-b border-dotted last:border-0 border-neutral-400
+						flex items-center
+					">
+						<div class="w-full overflow-hidden text-ellipsis text-sm">
+							{file.file.name}
+						</div>
+						<button
+							class="ml-2"
+							on:click={() => {removeFile(index);}}
+						>
+							<Icon src={X} class="w-6 h-6" />
+						</button>
+					</div>
+				{/each}
 			</div>
-		{/each}
+		{/if}
 	</div>
 </div>
